[EventDispatcher] Add LazyEventDispatcher
